#32
- Componenten proberen te maken

- Dropdownmenu nog niet helemaal werkend
