+++ src/routes/layout.svelte
@@ -1,0 +1,3 @@
+<script>
+    import Nav from '$lib/Nav.svelte'
+</script>
